Moved base picking protocols to a separate file
